+++ platform/assets/js/app.js
@@ -35,6 +35,7 @@ import {installSharedProfile} from "./shared_profile.js"
 import {mountDiscussionWorkbench} from "./discussion_workbench.js"
 import {mountHelloStream} from "./hello_stream.js"
 import {mountFixForm} from "./fix_form.js"
+import {mountCardTopUp} from "./card_topup.js"
 import {mountAgentFunding, mountAgentSetup, mountReadinessCard} from "./webmcp/agent_setup.js"
 import topbar from "../vendor/topbar"
 
@@ -67,6 +68,7 @@ const offerPageWideSurfaces = () => {
   mountHelloStream()
   mountFixForm({fetch: window.fetch.bind(window), csrfToken})
   mountAgentFunding()
+  mountCardTopUp()
   hideBrokenSiteLogos()
   mountDiscussionWorkbench()